+++ test/FactoryTest.php
@@ -1,35 +1,34 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-inputfilter for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-inputfilter/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-inputfilter/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\InputFilter;
+namespace LaminasTest\InputFilter;
 
+use Laminas\Filter;
+use Laminas\InputFilter\Factory;
+use Laminas\InputFilter\Input;
+use Laminas\InputFilter\InputFilter;
+use Laminas\InputFilter\InputFilterPluginManager;
+use Laminas\ServiceManager;
+use Laminas\Validator;
 use PHPUnit_Framework_TestCase as TestCase;
-use Zend\Filter;
-use Zend\InputFilter\Factory;
-use Zend\InputFilter\Input;
-use Zend\InputFilter\InputFilter;
-use Zend\Validator;
-use Zend\InputFilter\InputFilterPluginManager;
-use Zend\ServiceManager;
 
 class FactoryTest extends TestCase
 {
     public function testFactoryComposesFilterChainByDefault()
     {
         $factory = new Factory();
-        $this->assertInstanceOf('Zend\Filter\FilterChain', $factory->getDefaultFilterChain());
+        $this->assertInstanceOf('Laminas\Filter\FilterChain', $factory->getDefaultFilterChain());
     }
 
     public function testFactoryComposesValidatorChainByDefault()
     {
         $factory = new Factory();
-        $this->assertInstanceOf('Zend\Validator\ValidatorChain', $factory->getDefaultValidatorChain());
+        $this->assertInstanceOf('Laminas\Validator\ValidatorChain', $factory->getDefaultValidatorChain());
     }
 
     public function testFactoryAllowsInjectingFilterChain()
@@ -58,7 +57,7 @@ class FactoryTest extends TestCase
         $input = $factory->createInput(array(
             'name' => 'foo',
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $inputFilterChain = $input->getFilterChain();
         $this->assertNotSame($filterChain, $inputFilterChain);
         $this->assertSame($pluginManager, $inputFilterChain->getPluginManager());
@@ -74,7 +73,7 @@ class FactoryTest extends TestCase
         $input = $factory->createInput(array(
             'name' => 'foo',
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $inputValidatorChain = $input->getValidatorChain();
         $this->assertNotSame($validatorChain, $inputValidatorChain);
         $this->assertSame($validatorPlugins, $inputValidatorChain->getPluginManager());
@@ -97,10 +96,10 @@ class FactoryTest extends TestCase
                 'name' => 'foo',
             ),
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputFilterInterface', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilterInterface', $inputFilter);
         $this->assertEquals(1, count($inputFilter));
         $input = $inputFilter->get('foo');
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $inputFilterChain    = $input->getFilterChain();
         $inputValidatorChain = $input->getValidatorChain();
         $this->assertSame($filterPlugins, $inputFilterChain->getPluginManager());
@@ -126,20 +125,20 @@ class FactoryTest extends TestCase
                 ),
             ),
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $this->assertEquals('foo', $input->getName());
         $chain = $input->getFilterChain();
         $index = 0;
         foreach ($chain as $filter) {
             switch ($index) {
                 case 0:
-                    $this->assertInstanceOf('Zend\Filter\StringTrim', $filter);
+                    $this->assertInstanceOf('Laminas\Filter\StringTrim', $filter);
                     break;
                 case 1:
                     $this->assertSame($htmlEntities, $filter);
                     break;
                 case 2:
-                    $this->assertInstanceOf('Zend\Filter\StringToLower', $filter);
+                    $this->assertInstanceOf('Laminas\Filter\StringToLower', $filter);
                     $this->assertEquals('ISO-8859-1', $filter->getEncoding());
                     break;
                 default:
@@ -169,20 +168,20 @@ class FactoryTest extends TestCase
                 ),
             ),
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $this->assertEquals('foo', $input->getName());
         $chain = $input->getValidatorChain();
         $index = 0;
         foreach ($chain as $validator) {
             switch ($index) {
                 case 0:
-                    $this->assertInstanceOf('Zend\Validator\NotEmpty', $validator);
+                    $this->assertInstanceOf('Laminas\Validator\NotEmpty', $validator);
                     break;
                 case 1:
                     $this->assertSame($digits, $validator);
                     break;
                 case 2:
-                    $this->assertInstanceOf('Zend\Validator\StringLength', $validator);
+                    $this->assertInstanceOf('Laminas\Validator\StringLength', $validator);
                     $this->assertEquals(3, $validator->getMin());
                     $this->assertEquals(5, $validator->getMax());
                     break;
@@ -201,7 +200,7 @@ class FactoryTest extends TestCase
             'required' => false,
             'allow_empty' => false,
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $this->assertFalse($input->isRequired());
         $this->assertFalse($input->allowEmpty());
     }
@@ -213,7 +212,7 @@ class FactoryTest extends TestCase
             'name'        => 'foo',
             'allow_empty' => true,
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $this->assertTrue($input->allowEmpty());
         $this->assertFalse($input->isRequired());
     }
@@ -224,7 +223,7 @@ class FactoryTest extends TestCase
         $input   = $factory->createInput(array(
             'name'        => 'foo',
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $this->assertEquals('foo', $input->getName());
     }
 
@@ -235,7 +234,7 @@ class FactoryTest extends TestCase
             'name'              => 'foo',
             'continue_if_empty' => true,
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $this->assertTrue($input->continueIfEmpty());
     }
 
@@ -248,7 +247,7 @@ class FactoryTest extends TestCase
             'foo' => $input
         ));
 
-        $this->assertInstanceOf('Zend\InputFilter\InputFilterInterface', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilterInterface', $inputFilter);
         $this->assertTrue($inputFilter->has('foo'));
         $this->assertEquals($input, $inputFilter->get('foo'));
     }
@@ -262,7 +261,7 @@ class FactoryTest extends TestCase
             'foo' => $input
         ));
 
-        $this->assertInstanceOf('Zend\InputFilter\InputFilterInterface', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilterInterface', $inputFilter);
         $this->assertTrue($inputFilter->has('foo'));
         $this->assertEquals($input, $inputFilter->get('foo'));
     }
@@ -302,7 +301,7 @@ class FactoryTest extends TestCase
                 ),
             ),
             'baz' => array(
-                'type'   => 'Zend\InputFilter\InputFilter',
+                'type'   => 'Laminas\InputFilter\InputFilter',
                 'foo' => array(
                     'name'       => 'foo',
                     'required'   => false,
@@ -335,7 +334,7 @@ class FactoryTest extends TestCase
                 ),
             ),
             'bat' => array(
-                'type' => 'ZendTest\InputFilter\TestAsset\CustomInput',
+                'type' => 'LaminasTest\InputFilter\TestAsset\CustomInput',
                 'name' => 'bat',
             ),
             'zomg' => array(
@@ -343,7 +342,7 @@ class FactoryTest extends TestCase
                 'continue_if_empty' => true,
             ),
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputFilter', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilter', $inputFilter);
         $this->assertEquals(5, count($inputFilter));
 
         foreach (array('foo', 'bar', 'baz', 'bat', 'zomg') as $name) {
@@ -351,33 +350,33 @@ class FactoryTest extends TestCase
 
             switch ($name) {
                 case 'foo':
-                    $this->assertInstanceOf('Zend\InputFilter\Input', $input);
+                    $this->assertInstanceOf('Laminas\InputFilter\Input', $input);
                     $this->assertFalse($input->isRequired());
                     $this->assertEquals(2, count($input->getValidatorChain()));
                     break;
                 case 'bar':
-                    $this->assertInstanceOf('Zend\InputFilter\Input', $input);
+                    $this->assertInstanceOf('Laminas\InputFilter\Input', $input);
                     $this->assertTrue($input->allowEmpty());
                     $this->assertEquals(2, count($input->getFilterChain()));
                     break;
                 case 'baz':
-                    $this->assertInstanceOf('Zend\InputFilter\InputFilter', $input);
+                    $this->assertInstanceOf('Laminas\InputFilter\InputFilter', $input);
                     $this->assertEquals(2, count($input));
                     $foo = $input->get('foo');
-                    $this->assertInstanceOf('Zend\InputFilter\Input', $foo);
+                    $this->assertInstanceOf('Laminas\InputFilter\Input', $foo);
                     $this->assertFalse($foo->isRequired());
                     $this->assertEquals(2, count($foo->getValidatorChain()));
                     $bar = $input->get('bar');
-                    $this->assertInstanceOf('Zend\InputFilter\Input', $bar);
+                    $this->assertInstanceOf('Laminas\InputFilter\Input', $bar);
                     $this->assertTrue($bar->allowEmpty());
                     $this->assertEquals(2, count($bar->getFilterChain()));
                     break;
                 case 'bat':
-                    $this->assertInstanceOf('ZendTest\InputFilter\TestAsset\CustomInput', $input);
+                    $this->assertInstanceOf('LaminasTest\InputFilter\TestAsset\CustomInput', $input);
                     $this->assertEquals('bat', $input->getName());
                     break;
                 case 'zomg':
-                    $this->assertInstanceOf('Zend\InputFilter\Input', $input);
+                    $this->assertInstanceOf('Laminas\InputFilter\Input', $input);
                     $this->assertTrue($input->continueIfEmpty());
             }
         }
@@ -391,7 +390,7 @@ class FactoryTest extends TestCase
         ));
 
         $this->assertTrue($inputFilter->has('foo'));
-        $this->assertInstanceOf('Zend\InputFilter\Input', $inputFilter->get('foo'));
+        $this->assertInstanceOf('Laminas\InputFilter\Input', $inputFilter->get('foo'));
     }
 
     public function testFactoryAllowsPassingValidatorChainsInInputSpec()
@@ -402,7 +401,7 @@ class FactoryTest extends TestCase
             'name'       => 'foo',
             'validators' => $chain,
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $test = $input->getValidatorChain();
         $this->assertSame($chain, $test);
     }
@@ -415,7 +414,7 @@ class FactoryTest extends TestCase
             'name'    => 'foo',
             'filters' => $chain,
         ));
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
         $test = $input->getFilterChain();
         $this->assertSame($chain, $test);
     }
@@ -425,14 +424,14 @@ class FactoryTest extends TestCase
         $factory = new Factory();
 
         $inputFilter = $factory->createInputFilter(array(
-            'type'        => 'Zend\InputFilter\CollectionInputFilter',
+            'type'        => 'Laminas\InputFilter\CollectionInputFilter',
             'required'    => true,
             'inputfilter' => new InputFilter(),
             'count'       => 3,
         ));
 
-        $this->assertInstanceOf('Zend\InputFilter\CollectionInputFilter', $inputFilter);
-        $this->assertInstanceOf('Zend\InputFilter\InputFilter', $inputFilter->getInputFilter());
+        $this->assertInstanceOf('Laminas\InputFilter\CollectionInputFilter', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilter', $inputFilter->getInputFilter());
         $this->assertTrue($inputFilter->getIsRequired());
         $this->assertEquals(3, $inputFilter->getCount());
     }
@@ -456,11 +455,11 @@ class FactoryTest extends TestCase
             'filters' => array(
                 array(
                     'name'      => 'string_trim',
-                    'priority'  => \Zend\Filter\FilterChain::DEFAULT_PRIORITY - 1 // 999
+                    'priority'  => \Laminas\Filter\FilterChain::DEFAULT_PRIORITY - 1 // 999
                 ),
                 array(
                     'name'      => 'string_to_upper',
-                    'priority'  => \Zend\Filter\FilterChain::DEFAULT_PRIORITY + 1 //1001
+                    'priority'  => \Laminas\Filter\FilterChain::DEFAULT_PRIORITY + 1 //1001
                 ),
                 array(
                     'name'      => 'string_to_lower', // default priority 1000
@@ -477,13 +476,13 @@ class FactoryTest extends TestCase
         foreach ($input->getFilterChain()->getFilters() as $filter) {
             switch ($index) {
                 case 0:
-                    $this->assertInstanceOf('Zend\Filter\StringToUpper', $filter);
+                    $this->assertInstanceOf('Laminas\Filter\StringToUpper', $filter);
                     break;
                 case 1:
-                    $this->assertInstanceOf('Zend\Filter\StringToLower', $filter);
+                    $this->assertInstanceOf('Laminas\Filter\StringToLower', $filter);
                     break;
                 case 2:
-                    $this->assertInstanceOf('Zend\Filter\StringTrim', $filter);
+                    $this->assertInstanceOf('Laminas\Filter\StringTrim', $filter);
                     break;
             }
             $index++;
@@ -502,7 +501,7 @@ class FactoryTest extends TestCase
             )
         );
 
-        $this->assertInstanceOf('Zend\InputFilter\InputFilter', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilter', $inputFilter);
         $this->assertTrue($inputFilter->has('type'));
     }
 
@@ -513,7 +512,7 @@ class FactoryTest extends TestCase
             'type'        => 'collection',
             'input_filter' => new InputFilter(),
         ));
-        $this->assertInstanceOf('ZendTest\InputFilter\TestAsset\CustomFactory', $inputFilter->getFactory());
+        $this->assertInstanceOf('LaminasTest\InputFilter\TestAsset\CustomFactory', $inputFilter->getFactory());
     }
 
     /**
@@ -524,7 +523,7 @@ class FactoryTest extends TestCase
         $factory = new Factory();
         $input   = $factory->createInput(array(
             'name'          => 'test',
-            'type'          => 'Zend\InputFilter\Input',
+            'type'          => 'Laminas\InputFilter\Input',
             'error_message' => 'Custom error message',
         ));
         $this->assertEquals('Custom error message', $input->getErrorMessage());
@@ -540,11 +539,11 @@ class FactoryTest extends TestCase
         $factory = new Factory();
         $factory->setInputFilterManager($inputFilterManager);
         $this->assertInstanceOf(
-            'Zend\Validator\ValidatorPluginManager',
+            'Laminas\Validator\ValidatorPluginManager',
             $factory->getDefaultValidatorChain()->getPluginManager()
         );
         $this->assertInstanceOf(
-            'Zend\Filter\FilterPluginManager',
+            'Laminas\Filter\FilterPluginManager',
             $factory->getDefaultFilterChain()->getPluginManager()
         );
     }
@@ -567,7 +566,7 @@ class FactoryTest extends TestCase
     /**
      * @group 5691
      *
-     * @covers \Zend\InputFilter\Factory::createInput
+     * @covers \Laminas\InputFilter\Factory::createInput
      */
     public function testSetsBreakChainOnFailure()
     {
@@ -592,7 +591,7 @@ class FactoryTest extends TestCase
             ),
         ));
 
-        $this->assertInstanceOf('Zend\InputFilter\InputFilter', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilter', $inputFilter);
         $this->assertEquals(2, count($inputFilter));
         $this->assertTrue($inputFilter->has('foo'));
         $this->assertFalse($inputFilter->has('bar'));
@@ -604,8 +603,8 @@ class FactoryTest extends TestCase
      */
     public function testCanCreateInputFromProvider()
     {
-        /* @group $provider \Zend\InputFilter\InputProviderInterface|\PHPUnit_Framework_MockObject_MockObject */
-        $provider = $this->getMock('Zend\InputFilter\InputProviderInterface', array('getInputSpecification'));
+        /* @group $provider \Laminas\InputFilter\InputProviderInterface|\PHPUnit_Framework_MockObject_MockObject */
+        $provider = $this->getMock('Laminas\InputFilter\InputProviderInterface', array('getInputSpecification'));
 
         $provider
             ->expects($this->any())
@@ -615,7 +614,7 @@ class FactoryTest extends TestCase
         $factory = new Factory();
         $input   = $factory->createInput($provider);
 
-        $this->assertInstanceOf('Zend\InputFilter\InputInterface', $input);
+        $this->assertInstanceOf('Laminas\InputFilter\InputInterface', $input);
     }
 
     /**
@@ -623,9 +622,9 @@ class FactoryTest extends TestCase
      */
     public function testCanCreateInputFilterFromProvider()
     {
-        /* @group $provider \Zend\InputFilter\InputFilterProviderInterface|\PHPUnit_Framework_MockObject_MockObject */
+        /* @group $provider \Laminas\InputFilter\InputFilterProviderInterface|\PHPUnit_Framework_MockObject_MockObject */
         $provider = $this->getMock(
-            'Zend\InputFilter\InputFilterProviderInterface',
+            'Laminas\InputFilter\InputFilterProviderInterface',
             array('getInputFilterSpecification')
         );
         $provider
@@ -645,6 +644,6 @@ class FactoryTest extends TestCase
         $factory     = new Factory();
         $inputFilter = $factory->createInputFilter($provider);
 
-        $this->assertInstanceOf('Zend\InputFilter\InputFilterInterface', $inputFilter);
+        $this->assertInstanceOf('Laminas\InputFilter\InputFilterInterface', $inputFilter);
     }
 }